+++ push-service/src/main/java/com/skb/pushservice/config/MessageConsumerConfig.java
@@ -1,6 +1,6 @@
 package com.skb.pushservice.config;
 
-import com.skb.pushservice.dto.ExistDto;
+import com.skb.pushservice.domain.Exist.ExistDto;
 import org.apache.kafka.clients.consumer.ConsumerConfig;
 import org.apache.kafka.common.serialization.StringDeserializer;
 import org.springframework.beans.factory.annotation.Value;
