+++ push-service/src/main/java/com/skb/pushservice/domain/Exist/ExistDto.java
@@ -1,7 +1,8 @@
-package com.skb.pushservice.dto;
+package com.skb.pushservice.domain.Exist;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import com.skb.pushservice.domain.WatchInfo.WatchInfoDto;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
